Allow modifying the SPRT bounds during a test
This is useful if you started a test as a progression test, but realise
that the Elo gain isn't as high as you'd want. Nevertheless, you're
happy to still merge the change if it's not a regression (e.g. it's
still a nice simplification) but don't want to re-do the full test.